backend/requirements.txt
asgiref==3.7.2
Django==3.2.3
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
python-dotenv==1.0.0
djoser==2.1.0
gunicorn==20.1.0
djangorestframework==3.12.4
